+++ app.config.ts
@@ -21,7 +21,7 @@ export default defineConfig({
 }).addRouter({
   name: "websocket",
   type: "http",
-  handler: "./app/server/ws.ts",
+  handler: "./lib/server/ws.ts",
   target: "server",
   base: "/_ws",
   plugins: () => [tsConfigPaths()],
