+++ app/routes/dashboard.tsx
@@ -1,5 +1,5 @@
 import { Link, Outlet, createFileRoute, redirect } from "@tanstack/react-router";
-import { Button } from "~/components/ui/button";
+import { Button } from "~/lib/components/ui/button";
 
 export const Route = createFileRoute("/dashboard")({
   component: DashboardLayout,
