+++ eslint.config.js
@@ -7,7 +7,6 @@ import reactHooks from "eslint-plugin-react-hooks";
 import globals from "globals";
 import tseslint from "typescript-eslint";
 
-// TODO: clean up for better composability
 export default tseslint.config(
   {
     ignores: ["dist", ".vinxi", ".wrangler", ".vercel", ".netlify", ".output", "build/"],
