+++ lib/server/auth.ts
